+++ packages/gasket-plugin-elastic-apm/lib/index.d.ts
@@ -1,5 +1,5 @@
 import type { IncomingMessage, ServerResponse } from 'http';
-import type { AgentConfigOptions } from 'elastic-apm-node';
+import type { AgentConfigOptions, Transaction } from 'elastic-apm-node';
 
 declare module '@gasket/engine' {
   export interface GasketConfig {
@@ -8,19 +8,13 @@ declare module '@gasket/engine' {
     },
   }
 
-  type RequestDetails = {
-    req: IncomingMessage,
-    res: ServerResponse
-  }
-
   export interface HookExecTypes {
-    transactionName(
-      currentName: string,
-      details: RequestDetails
-    ): MaybeAsync<string>,
-    transactionLabels(
-      currentLabels: Record<string, string>,
-      details: RequestDetails
-    ): MaybeAsync<Record<string, string>>,
+    apmTransaction(
+      transaction: Transaction,
+      details: {
+        req: IncomingMessage,
+        res: ServerResponse
+      }
+    ): MaybeAsync<void>
   }
 }