+++ packages/gasket-plugin-elastic-apm/lib/middleware.js
@@ -1,5 +1,6 @@
 /* eslint-disable spaced-comment */
 // @ts-check
+/// <reference types="./index" />
 /// <reference types="../../gasket-plugin-nextjs" />
 
 /**
@@ -12,29 +13,21 @@ const { callbackify } = require('util');
 const apm = require('elastic-apm-node');
 
 /**
- * Middleware which renames transactions
+ * Middleware for customizing transactions
  *
  * @param {Gasket}    gasket  The Gasket engine
  * @param {Request}   req     The HTTP request being handled
  * @param {Response}  res     The server response
  */
-async function transactionRenamingMiddleware(gasket, req, res) {
+async function customizeTransaction(gasket, req, res) {
   const transaction = apm.currentTransaction;
   if (!transaction) {
     return;
   }
 
-  const requestDetails = { req, res };
-
-  const [name, labels] = await Promise.all([
-    gasket.execWaterfall('transactionName', transaction.name, requestDetails),
-    gasket.execWaterfall('transactionLabels', {}, requestDetails)
-  ]);
-
-  transaction.name = name;
-  transaction.addLabels(labels);
+  await gasket.exec('apmTransaction', transaction, { req, res });
 }
 
 module.exports = (gasket) => [
-  callbackify(async (req, res) => transactionRenamingMiddleware(gasket, req, res))
+  callbackify(async (req, res) => customizeTransaction(gasket, req, res))
 ];