+++ packages/gasket-plugin-nextjs/lib/transaction-name.js
@@ -1,4 +1,0 @@
-module.exports = async (gasket, name, { req }) => {
-  const route = await req.getNextRoute();
-  return route ? route.page : name;
-};