+++ packages/gasket-plugin-nextjs/test/index.test.js
@@ -21,14 +21,13 @@ describe('Plugin', function () {
 
   it('has expected hooks', () => {
     const expected = [
+      'apmTransaction',
       'build',
       'configure',
       'create',
       'express',
       'metadata',
       'middleware',
-      'transactionLabels',
-      'transactionName',
       'workbox'
     ];
 
@@ -173,24 +172,6 @@ describe('express hook', () => {
 
     const nextOptions = next.lastCall.args[0];
     assume(nextOptions.conf).to.not.haveOwnProperty('webpack');
-  });
-
-  it('executes nextPreHandling before next.js handles a request', async () => {
-    const gasket = mockGasketApi();
-    await hook(gasket, expressApp, false);
-
-    const routeHandler = expressApp.all.lastCall.args[1];
-
-    const mockReq = { headers: {} };
-    const mockRes = { locals: { gasketData: { } } };
-    const mockNext = stub();
-
-    await routeHandler(mockReq, mockRes, mockNext);
-    assume(gasket.exec).has.been.calledWithMatch('nextPreHandling', {
-      req: mockReq,
-      res: mockRes,
-      next: nextHandler
-    });
   });
 });